WIP: A proof-of-concept showing the adoption of flakes with agenix
This commit turns this repository into a flake, and obtains
dependencies (such as stable and unstable versions of nixpkgs) through
flakes, rather than by direct references to tarballs.

In order to allow secrets to be committed to this repository (which is
necessary, as flakes are not designed to depend on files outside of
flakes), this commit depends on the agenix [1] nixos module, which
will decrypt and mount age-encrypted [2] files as part of system
activation. This should allow all necessary secrets, and the nix files
which reference them, to be committed to this repository without
danger of leaking sensitive information.

Note that much of the contents of the "secrets" folder are
placeholders, and should be replaced with actual public keys before
attempting to deploy.

1. https://github.com/ryantm/agenix
2. https://github.com/FiloSottile/age